Reintegrate: mask Eiger uint32 dummies, 2D-only, reliable Stop popup
The reintegrate route diverged from the (clean) fresh integrate in three ways,
all fixed here:

* Saturation spike: the live wrangler stamps a bad-pixel mask on every
  LiveFrame (negatives + the uint32 dead/hot sentinel 4294967295, e.g. Eiger
  masters), but a frame lazy-loaded from the .nxs for reintegration carries
  mask=None (the per-frame mask is not persisted).  On real Eiger data that left
  250,588 dummy pixels at 4.29e9 counts unmasked, dominating every radial bin
  (1D max 4.295e9 -> the high-Q spike).  `Mask Saturated` couldn't help: the
  uint32 ceiling is excluded from saturation_pixels by design.  Fix: a single
  shared `compute_bad_pixel_mask` (xdart.modules.reduction), reattached in
  `_prepare_frame_for_headless_reduction`, recomputed per pass so a toggle change
  between Reintegrate clicks is honoured.  `_lazy_load_raw` also keeps the source
  integer dtype so the ceiling survives to the reduce.

* 2D reintegrate recomputed the 1D: `integrate_1d = not integrate_2d` so each
  Reintegrate button does only its own dimension.  Recomputing the 1D overwrote
  the good curve and forced a 1D-stack rewrite that crashed a partial save.

* Stop "discard?" popup never showed: the pre-check read np.shape() of an
  IntegrationResult dataclass (-> () -> [-1] IndexError, swallowed).  Now reads
  .intensity, plus an authoritative post-first-frame axis/unit/shape comparison
  so the popup fires for any reason the writer would reject a partial rewrite.

Also: reduce only the dimension the plan computed (preserve int_1d on a 2D-only
pass); single per-frame `update` emit (was doubled); `[REINT-MASK]` diagnostics
log what each pass rejects.  Full suite + GI equivalence spine green.

+++ src/xdart/modules/ewald/frame.py
@@ -401,18 +401,25 @@ class LiveFrame():
             if ext in (".h5", ".nxs", ".hdf5"):
                 from xrd_tools.io.nexus import open_nexus_image_stack
                 with open_nexus_image_stack(full) as stack:
-                    self.map_raw = np.asarray(
-                        stack[fidx], dtype=np.float32,
-                    )
+                    # Keep the source INTEGER dtype (Eiger uint16/uint32).  The
+                    # reduce derives the Mask-Saturated ceiling from the integer
+                    # dtype (integer_saturation_ceiling -> None for float) and
+                    # only then converts to float — casting to float32 HERE lost
+                    # the ceiling, silently disabling saturation masking on a
+                    # reintegrate (the high-Q edge spike vs a clean fresh
+                    # integrate, identical settings).  Consumers (display, stitch,
+                    # reduce) all np.asarray(..., float) as needed.
+                    self.map_raw = np.asarray(stack[fidx])
             else:
                 # O2: forward source_frame_idx for multi-frame
                 # non-HDF5 sources too (e.g. multi-frame TIFF /
                 # CBF stacks).  Single-frame files use frame=0
                 # and ignore the kwarg, so this is safe.
                 from xrd_tools.io.image import read_image
-                self.map_raw = np.asarray(
-                    read_image(full, frame=fidx), dtype=np.float32,
-                )
+                # Keep the source dtype (see the HDF5 branch above): integer raws
+                # must stay integer so the saturation ceiling survives to the
+                # reduce; consumers convert to float as needed.
+                self.map_raw = np.asarray(read_image(full, frame=fidx))
         except Exception as e:
             import logging
             logging.getLogger(__name__).debug(
